MGMT-797: Avoid storing duplicate images
Change the ID of the ISO that we generate to be deterministic based on
the cluster ID and the parameters passed so that we don't store multiple
identical images in S3. In case the image already exists we will still
generate and upload, but it will replace the existing one and not incur
additional storage fees. In the future we can optimize avoiding the
generation and upload.