Added Monitoring integration for Assisted Installer.
This feature will involve 2 flows, the first one for Minikube or K8s
which will deploy OLM, then Prometheus and in the end Grafana.

The other flow consist on the deployment directly of Prometheus
integrated with oauth-proxy as a sidecar and the second part is grafana
which is also integrated with oauth-proxy.

Reported-by: Rom Freiman
Signed-off-by: Juan Manuel Parrilla <jparrill@redhat.com>