MGMT-948: Change ClusterUpdate and GenerateClusterISO to overwrite all values
We cannot determine which values are passed as empty/default and which
are not passed at all. This means that once a user sets a value for a
property, they cannot set it back to the default. We now change the
semantics of these two APIs such that the property for any value not
passed is set to the default.