Subsystem: Added cancel installation tests
Added cancel installation test which starts an installation, then calling
to the cancel method and verifies that the cluster & hosts states are set
to error. Second test will try to cancel installation before it began and
except to get failure. Third test will try to cancel when the cluster is
already in error state, and verify that the host sates have also changed.
Forth will try to cancel hosts in multiple states, expected to fail.