MGMT-1526 Host: clear bootstrap field on setRole to Worker
Since we decided to reuse a bootstrap node we don't clear this field on reset.
A bootstrap's role can be set to Worker after reset.  Currently nothing clears
this field. If we install and reset in the second time, then return the former
bootstrap back to Mater role, we will have two bootstraps. In this commit the
bootstrap field is zeroed when setting a host role to Worker. Subsystem test is
included.