MGMT-1667: Use minikube profile per namespace
We want to be able to deploy more then one environment on the same host,
using minikube. In order to support initializing and destroying of a deployment
without effecting the rest of the environments we will use minikube profile
per deployment. By default, if no profile is specified, deploy services to
default 'minikube' namespace.