OCPBUGSM-13597 Update register transition
There are some edge cases on installation where user tries to reset
on the same time reboot is called. On some cases the agent will get
to reset itself and register again just before the reboot, before
the cluster monitor gets to set the status in
resetting-pending-user-action. In this commit I added to the host
statemachine a tranistion verifies that host can't register from
resetting state if the stage is reboot.